Rename App container AppContainer and fix imports
